refactored arrangeThesauri
